Item wear out
When using a Tool, a random damage is applied to its value,
when reaches 0, it'll break.

It's unnecessary to use double damage values, we don't really need
such accuracy, int16_t is totally fine.

+++ config/level.c
@@ -4,7 +4,7 @@
 /**
  * MATERIAL
  */
-double MATERIAL_STRENGTH[MATERIAL_NUM] = {
+int16_t MATERIAL_STRENGTH[MATERIAL_NUM] = {
     [SAND] = 50,
     [DIRT] = 80,
     [WOOD] = 100,
